+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic `Plugin` trait and `PingPlugin` implementation added to `plugins` crate.
 - Basic Jetpack Compose UI implemented in Android project (FFI simulated).
 - Basic auto-acceptance pairing implemented in `daemon` crate.
+- Basic plugin dispatcher and `PingPlugin` integrated into `daemon` crate.
 
 ### Changed
 - mDNS integration in `daemon` crate temporarily put on hold due to library issues.